+++ termpair/frontend_src/src/ErrorBoundary.tsx
@@ -1,0 +1,29 @@
+import React from "react";
+
+export class ErrorBoundary extends React.Component<any, any> {
+  constructor(props: any) {
+    super(props);
+    this.state = { hasError: false };
+  }
+
+  static getDerivedStateFromError(error: any) {
+    // Update state so the next render will show the fallback UI.
+    return { hasError: true };
+  }
+
+  componentDidCatch(error: any, errorInfo: any) {
+    // You can also log the error to an error reporting service
+    // logErrorToMyService(error, errorInfo);
+    console.error(error);
+    console.error(errorInfo);
+  }
+
+  render() {
+    if (this.state.hasError) {
+      // You can render any custom fallback UI
+      return <h1 className="text-white">Something went wrong.</h1>;
+    }
+
+    return this.props.children;
+  }
+}
